@@ -172,10 +189,9 @@
     "cross-spawn": "^6.0.5",
     "css-loader": "^2.1.1",
     "detect-port": "^1.3.0",
-    "electron": "^4.1.3",
+    "electron": "^8.2.4",
     "electron-builder": "^21.2.0",
     "electron-devtools-installer": "^2.2.4",
-    "electron-rebuild": "^1.11.0",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.11.2",
     "enzyme-to-json": "^3.3.5",
