@@ -6,14 +6,17 @@ import {
   WalletBackend,
   LogLevel,
   prettyPrintAmount,
-  WalletErrorCode,
-  FeeType
+  WalletErrorCode
 } from 'turtlecoin-wallet-backend';
 import log from 'electron-log';
 import { ipcRenderer } from 'electron';
 import { createObjectCsvWriter } from 'csv-writer';
 import { atomicToHuman, convertTimestamp } from '../mainWindow/utils/utils';
-import Configuration from '../Configure';
+import Configure from '../Configure';
+
+export function delay(ms: number) {
+    return new Promise((resolve) => setTimeout(resolve, ms));
+}
 
 export default class Backend {
   notifications: boolean;
@@ -74,7 +77,21 @@ export default class Backend {
   }
 
   getNodeFee(): void {
-    ipcRenderer.send('fromBackend', 'nodeFee', this.wallet.getNodeFee()[1]);
+    this.send('nodeFee', this.wallet.getNodeFee()[1]);
+  }
+
+  send(type: string, data: any) {
+    if (typeof data === 'object') {
+      // eslint-disable-next-line no-restricted-syntax
+      for (const key in data) {
+        if (typeof data[key] === 'function') {
+          // eslint-disable-next-line no-param-reassign
+          delete data[key];
+        }
+      }
+    }
+    log.debug({ type, data });
+    ipcRenderer.send('fromBackend', type, data);
   }
 
   getWalletActive(): boolean {
@@ -114,7 +131,7 @@ export default class Backend {
     const rawTransactions = this.getFormattedTransactions(
       undefined,
       undefined,
-      false
+      true
     );
     const csvWriter = createObjectCsvWriter({
       path: savePath,
@@ -144,23 +161,32 @@ export default class Backend {
     this.wallet.scanCoinbaseTransactions(value);
   }
 
+  setenableAutoOptimization(value: boolean) {
+    this.wallet.enableAutoOptimization(value);
+  }
+
   async sendTransaction(hash: string): void {
+    /* Wait for UI to load before blocking thread */
+    await delay(500);
+
     const result = await this.wallet.sendPreparedTransaction(hash);
 
     if (result.success) {
       console.log(
         `Sent transaction, hash ${
           result.transactionHash
-        }, fee ${prettyPrintAmount(result.fee)}`
+        }, fee ${prettyPrintAmount(result.fee, Configure)}`
       );
       const response = {
         status: 'SUCCESS',
         hash: result.transactionHash,
         error: undefined
       };
-      ipcRenderer.send('fromBackend', 'sendTransactionResponse', response);
-      this.getTransactions(this.getLastTxAmountRequested() + 1);
+      this.send('sendTransactionResponse', response);
+      await this.getTransactions(this.getLastTxAmountRequested() + 1);
     } else {
+      /* TODO: Optionally allow retries in case of network error? */
+      this.wallet.deletePreparedTransaction(hash);
       console.log(`Failed to send transaction: ${result.error.toString()}`);
       result.error.errorString = result.error.toString();
       const response = {
@@ -168,19 +194,39 @@ export default class Backend {
         hash: undefined,
         error: result.error
       };
-      ipcRenderer.send('fromBackend', 'sendTransactionResponse', response);
+      this.send('sendTransactionResponse', response);
     }
   }
 
-  async prepareTransaction(transaction): void {
+  async prepareTransaction(transaction): Promise<void> {
+    const [unlockedBalance, lockedBalance] = await this.wallet.getBalance();
+
+    const networkHeight: number = this.daemon.getNetworkBlockCount();
+
+    const [feeAddress, nodeFee] = this.wallet.getNodeFee();
+
+    let txFee = Configure.minimumFee;
+
     const { address, amount, paymentID, sendAll } = transaction;
 
-    const destinations = [[address, sendAll ? 100000 : amount]];
+    const payments = [];
+
+    if (sendAll) {
+        payments.push([
+            address,
+            (networkHeight >= Configure.feePerByteHeight) ? 1 : (unlockedBalance - nodeFee - txFee), /* Amount does not matter for sendAll destination */
+        ]);
+    } else {
+        payments.push([
+            address,
+            amount,
+        ]);
+    }
 
     const result = await this.wallet.sendTransactionAdvanced(
-      destinations, // destinations
+      payments, // destinations
       undefined, // mixin
-      FeeType.FixedFee(Configuration.minimumFee), // fee
+      (networkHeight >= Configure.feePerByteHeight) ? undefined : {isFixedFee: true, fixedFee: txFee}, // fee
       paymentID, // paymentID
       undefined, // subwalletsToTakeFrom
       undefined, // changeAddress
@@ -191,20 +237,37 @@ export default class Backend {
     log.info(result);
 
     if (result.success) {
-      const [unlockedBalance, lockedBalance] = this.wallet.getBalance();
-      const balance = parseInt(unlockedBalance + lockedBalance, 10);
+      let actualAmount = amount;
+
+      if (networkHeight >= Configure.feePerByteHeight) {
+        txFee = result.fee;
+      }
+
+      if (sendAll) {
+        let transactionSum = 0;
+
+        /* We could just get the sum by calling getBalance.. but it's
+        * possibly just changed. Safest to iterate over prepared
+        * transaction and calculate it. */
+        for (const input of result.preparedTransaction.inputs) {
+          transactionSum += input.input.amount;
+        }
+        actualAmount = transactionSum
+                     - txFee
+                     - nodeFee;
+      }
       const response = {
         status: 'SUCCESS',
         hash: result.transactionHash,
         address,
         paymentID,
-        amount: sendAll ? balance : amount,
-        fee: result.fee,
-        nodeFee: this.wallet.getNodeFee()[1],
+        amount: actualAmount,
+        fee: txFee,
+        nodeFee: nodeFee,
         error: undefined
       };
-      ipcRenderer.send('fromBackend', 'prepareTransactionResponse', response);
-      this.getTransactions(this.getLastTxAmountRequested() + 1);
+      this.send('prepareTransactionResponse', response);
+      await this.getTransactions(this.getLastTxAmountRequested() + 1);
     } else {
       console.log(`Failed to send transaction: ${result.error.toString()}`);
       result.error.errorString = result.error.toString();
@@ -214,20 +277,16 @@ export default class Backend {
         address,
         paymentID,
         amount,
-        fee: result.fee,
-        nodeFee: this.wallet.getNodeFee()[1],
+        fee: txFee,
+        nodeFee: nodeFee,
         error: result.error
       };
-      ipcRenderer.send('fromBackend', 'prepareTransactionResponse', response);
+      this.send('prepareTransactionResponse', response);
     }
   }
 
   verifyPassword(password: string): void {
-    ipcRenderer.send(
-      'fromBackend',
-      'authenticationStatus',
-      password === this.walletPassword
-    );
+    this.send('authenticationStatus', password === this.walletPassword);
   }
 
   changePassword(passwords: any): void {
@@ -244,27 +303,30 @@ export default class Backend {
         response = { status: 'FAILURE', error: 'SAVEERROR' };
       }
     }
-    ipcRenderer.send('fromBackend', 'passwordChangeResponse', response);
+    this.send('passwordChangeResponse', response);
   }
 
   async rescanWallet(height: number) {
     await this.wallet.reset(height);
     this.saveWallet(false);
-    ipcRenderer.send('fromBackend', 'rescanResponse', height);
+    this.send('rescanResponse', height);
   }
 
-  getFormattedTransactions(
+  async getFormattedTransactions(
     startIndex?: number,
     numTransactions?: number,
     includeFusions?: boolean
   ): any[] {
-    const rawTransactions = this.wallet.getTransactions(
+    const rawTransactions = await this.wallet.getTransactions(
       startIndex,
       numTransactions,
       includeFusions || false
     );
-    const [unlockedBalance, lockedBalance] = this.wallet.getBalance();
+
+    const [unlockedBalance, lockedBalance] = await this.wallet.getBalance();
+
     let balance = parseInt(unlockedBalance + lockedBalance, 10);
+
     const transactions = [];
 
     for (const [index, tx] of rawTransactions.entries()) {
@@ -295,21 +357,22 @@ export default class Backend {
     }
   }
 
-  getTransactions(displayCount: number): void {
+  async getTransactions(displayCount: number): void {
     this.setLastTxAmountRequested(displayCount);
-    ipcRenderer.send(
-      'fromBackend',
+    const get_tx = await this.getFormattedTransactions(0, displayCount, true);
+    this.send(
       'transactionList',
-      this.getFormattedTransactions(0, displayCount, false)
+      get_tx
     );
   }
 
   getTransactionCount(): void {
-    ipcRenderer.send('fromBackend', 'transactionCount', this.transactionCount);
+    this.send('transactionCount', this.transactionCount);
   }
 
-  getBalance(): void {
-    ipcRenderer.send('fromBackend', 'balance', this.wallet.getBalance());
+  async getBalance(): void {
+    console.log(`balance: ${this.wallet.getBalance()}`);
+    this.send('balance', await this.wallet.getBalance());
   }
 
   saveWallet(notify: boolean, path?: string): boolean {
@@ -322,28 +385,26 @@ export default class Backend {
     );
 
     if (notify) {
-      ipcRenderer.send('fromBackend', 'saveWalletResponse', status);
+      this.send('saveWalletResponse', status);
     }
     return status;
   }
 
-  transactionSearch(query: string) {
-    const transactions = this.wallet.getTransactions();
+  async transactionSearch(query: string) {
+    const transactions = await this.wallet.getTransactions();
     const possibleTransactionValues = ['blockHeight', 'hash', 'paymentID'];
     const transactionResults = possibleTransactionValues.map(value => {
       return this.search(query, transactions, value);
     });
+
     let sanitizedResults = [];
     /* the search function returns a separate array of results for each
     value searched, we need to concat them together with spread */
     for (let i = 0; i < transactionResults.length; i++) {
       sanitizedResults = [...transactionResults[i], ...sanitizedResults];
     }
-    ipcRenderer.send(
-      'fromBackend',
-      'transactionSearchResponse',
-      sanitizedResults
-    );
+
+    this.send('transactionSearchResponse', sanitizedResults);
   }
 
   search(searchedValue: any, arrayToSearch: any[], objectPropertyName: string) {
@@ -357,8 +418,17 @@ export default class Backend {
       ) {
         /* we have to disable this because the function gets lost
         when we send the object over ipc */
+
         // eslint-disable-next-line no-param-reassign
         arrayToSearch[i].totalTxAmount = arrayToSearch[i].totalAmount();
+
+        // we need to delete the function afterwards because of
+        // electron 9's new serialization code
+        // https://www.electronjs.org/docs/breaking-changes#behavior-changed-values-sent-over-ipc-are-now-serialized-with-structured-clone-algorithm
+
+        // eslint-disable-next-line no-param-reassign
+        delete arrayToSearch[i].totalAmount;
+
         resultsToReturn.push(arrayToSearch[i]);
       }
     }
@@ -374,11 +444,7 @@ export default class Backend {
   }
 
   getConnectionInfo(): void {
-    ipcRenderer.send(
-      'fromBackend',
-      'daemonConnectionInfo',
-      this.wallet.getDaemonConnectionInfo()
-    );
+    this.send('daemonConnectionInfo', this.wallet.getDaemonConnectionInfo());
   }
 
   setLogLevel(logLevel: string): void {
@@ -389,21 +455,24 @@ export default class Backend {
   async walletInit(wallet: any): Promise<void> {
     this.wallet = wallet;
     this.setLogLevel(this.logLevel);
-    this.wallet.enableAutoOptimization(false);
     this.wallet.on(
       'heightchange',
-      (walletBlockCount, localDaemonBlockCount, networkBlockCount) => {
-        ipcRenderer.send('fromBackend', 'syncStatus', [
+      async (walletBlockCount, localDaemonBlockCount, networkBlockCount) => {
+        this.send('syncStatus', [
           walletBlockCount,
           localDaemonBlockCount,
           networkBlockCount
         ]);
+        if (networkBlockCount !== 0 && walletBlockCount !== 0 && (networkBlockCount - walletBlockCount) < 40) {
+            this.send('balance', await this.wallet.getBalance());
+        }
       }
     );
-    this.wallet.on('transaction', () => {
+
+    this.wallet.on('transaction', async () => {
       this.getTransactionCount();
-      this.getTransactions(this.getLastTxAmountRequested() + 1);
-      this.getBalance();
+      await this.getTransactions(this.getLastTxAmountRequested() + 1);
+      await this.getBalance();
     });
 
     this.wallet.on('incomingtx', transaction => {
@@ -413,50 +482,45 @@ export default class Backend {
           body: `You've just received ${atomicToHuman(
             transaction.totalAmount(),
             true
-          )} TRTL.`
+          )} {Configure.ticker}.`
         });
       }
     });
     this.setWalletActive(true);
-    ipcRenderer.send('fromBackend', 'syncStatus', this.wallet.getSyncStatus());
-    ipcRenderer.send(
-      'fromBackend',
-      'primaryAddress',
-      this.wallet.getPrimaryAddress()
-    );
-    ipcRenderer.send(
-      'fromBackend',
-      'transactionList',
-      this.getFormattedTransactions(0, 50, false)
-    );
+    this.send('syncStatus', this.wallet.getSyncStatus());
+    this.send('primaryAddress', this.wallet.getPrimaryAddress());
+    this.send('transactionList', await this.getFormattedTransactions(0, 50, true));
     this.getTransactionCount();
-    ipcRenderer.send('fromBackend', 'balance', this.wallet.getBalance());
-    ipcRenderer.send('fromBackend', 'walletActiveStatus', true);
-    ipcRenderer.send('fromBackend', 'authenticationStatus', true);
+    this.send('balance', await this.wallet.getBalance());
+    this.send('walletActiveStatus', true);
+    this.send('authenticationStatus', true);
     await this.wallet.start();
-    ipcRenderer.send(
-      'fromBackend',
-      'daemonConnectionInfo',
-      this.wallet.getDaemonConnectionInfo()
-    );
+    //this.wallet.usingNativeCrypto();
+    this.send('daemonConnectionInfo', this.wallet.getDaemonConnectionInfo());
     this.getNodeFee();
   }
 
-  getSecret(): string {
+  async getMnemonic(): string {
+      let [mnemonicSeed, err] = await this.wallet.getMnemonicSeed();
+      if (err) {
+        if (err.errorCode === 41) {
+          mnemonicSeed = '';
+        } else {
+          throw err;
+        }
+      }
+      return mnemonicSeed;
+  }
+
+  async getSecret(): string {
     const publicAddress = this.wallet.getPrimaryAddress();
     const [
       privateSpendKey,
       privateViewKey
     ] = this.wallet.getPrimaryAddressPrivateKeys();
     // eslint-disable-next-line prefer-const
-    let [mnemonicSeed, err] = this.wallet.getMnemonicSeed();
-    if (err) {
-      if (err.errorCode === 41) {
-        mnemonicSeed = '';
-      } else {
-        throw err;
-      }
-    }
+
+    let mnemonicSeed = await this.getMnemonic();
 
     const secret =
       // eslint-disable-next-line prefer-template
@@ -472,21 +536,21 @@ export default class Backend {
     return secret;
   }
 
-  startWallet(password: string): void {
+  async startWallet(password: string): Promise<void> {
     this.walletPassword = password;
-    const [openWallet, error] = WalletBackend.openWalletFromFile(
+    const [openWallet, error] = await WalletBackend.openWalletFromFile(
       this.daemon,
       this.walletFile,
       this.walletPassword,
-      Configuration
+      Configure
     );
     if (!error) {
       this.walletInit(openWallet);
     } else if (error.errorCode === WalletErrorCode.WRONG_PASSWORD) {
-      ipcRenderer.send('fromBackend', 'authenticationStatus', false);
+      this.send('authenticationStatus', false);
     } else {
       error.errorString = error.toString();
-      ipcRenderer.send('fromBackend', 'authenticationError', error);
+      this.send('authenticationError', error);
     }
   }
 }
