@@ -84,7 +84,7 @@ export default class FirstStartup extends Component<Props, State> {
           <div className="mid-div">
             <div className={`box loginbox passwordchangebox ${fillColor}`}>
               <h1 className={`title has-text-centered ${redTitleColor}`}>
-                {il8n.welcome_to_proton}
+                {il8n.welcome_to_pluton}
               </h1>
               <button
                 className={`button is-large is-fullwidth ${buttonColor}`}
