@@ -5,13 +5,13 @@ import React, { Component } from 'react';
 import ReactTooltip from 'react-tooltip';
 import { remote } from 'electron';
 import log from 'electron-log';
-import { WalletBackend } from 'turtlecoin-wallet-backend';
+import { WalletBackend, Daemon } from 'turtlecoin-wallet-backend';
 import NavBar from './NavBar';
 import BottomBar from './BottomBar';
 import Redirector from './Redirector';
 import { uiType } from '../utils/utils';
 import { eventEmitter, reInitWallet, config } from '../index';
-import Configuration from '../../Configure';
+import Configure from '../../Configure';
 
 type State = {
   darkMode: boolean,
@@ -22,7 +22,7 @@ type State = {
   showPassword: boolean,
   darkMode: boolean,
   mnemonicSeed: string,
-  importedWallet: any,
+  importedWallet: WalletBackend | null,
   scanHeight: string
 };
 
@@ -116,7 +116,7 @@ export default class Import extends Component<Props, State> {
     });
   };
 
-  nextPage = () => {
+  nextPage = async () => {
     const {
       activePage,
       password,
@@ -130,11 +130,11 @@ export default class Import extends Component<Props, State> {
     let currentPageNumber: number = this.evaluatePageNumber(activePage);
 
     if (currentPageNumber === 1) {
-      const [restoredWallet, error] = WalletBackend.importWalletFromSeed(
-        Configuration.defaultDaemon,
+      const [restoredWallet, error] = await WalletBackend.importWalletFromSeed(
+        Configure.defaultDaemon,
         scanHeight === '' ? 0 : Number(scanHeight),
         mnemonicSeed,
-        Configuration
+        Configure
       );
 
       if (error) {
@@ -171,13 +171,17 @@ export default class Import extends Component<Props, State> {
           }
         ]
       };
-      const savePath = remote.dialog.showSaveDialog(null, options);
-      if (savePath === undefined) {
+      const response = await remote.dialog.showSaveDialog(null, options);
+      if (response.canceled) {
         return;
       }
-      const saved = importedWallet.saveWalletToFile(savePath, password);
+
+      const saved = importedWallet.saveWalletToFile(
+        `${response.filePath}`,
+        password
+      );
       if (saved) {
-        reInitWallet(savePath);
+        reInitWallet(`${response.filePath}`);
       } else {
         const message = (
           <div>
