@@ -6,6 +6,7 @@ import log from 'electron-log';
 import { DaemonConnection } from 'turtlecoin-wallet-backend';
 import { config, eventEmitter } from '../index';
 import { roundToNearestHundredth } from '../utils/utils';
+import Configure from '../../Configure';
 
 export default class WalletSession {
   loginFailed: boolean;
@@ -17,6 +18,8 @@ export default class WalletSession {
   selectedFiat: string;
 
   fiatPrice: number;
+
+  selectedTimeZone: string;
 
   primaryAddress: string;
 
@@ -45,8 +48,11 @@ export default class WalletSession {
     this.loginFailed = false;
     this.firstLoadOnLogin = true;
     this.selectedFiat = config.selectedFiat;
+    this.selectedTimeZone = config.selectedTimeZone;
     this.fiatPrice = 0;
     this.getFiatPrice(this.selectedFiat);
+    this.updateNodeList();
+    this.getTimeZone(this.selectedTimeZone);
   }
 
   setPreparedTransactionHash(hash: string): void {
@@ -112,7 +118,7 @@ export default class WalletSession {
   }
 
   getFiatPrice = async (fiat: string) => {
-    const apiURL = `https://api.coingecko.com/api/v3/coins/markets?vs_currency=${fiat}&ids=turtlecoin&order=market_cap_desc&per_page=100&page=1&sparkline=false&price_change_percentage=7d`;
+    const apiURL = `https://api.coingecko.com/api/v3/coins/markets?vs_currency=${fiat}&ids=${Configure.coinName.toLowerCase()}&order=market_cap_desc&per_page=100&page=1&sparkline=false&price_change_percentage=7d`;
 
     const requestOptions = {
       method: 'GET',
@@ -130,6 +136,45 @@ export default class WalletSession {
       log.debug(`Request failed, CoinGecko API call error: \n`, err);
       return undefined;
     }
+  };
+
+
+  updateNodeList = async () => {
+    const apiURL = `${Configure.nodeListURL}`;
+
+    const requestOptions = {
+      method: 'GET',
+      timeout: Configure.requestTimeout,
+      uri: apiURL,
+      headers: {},
+      json: true,
+      gzip: false
+    };
+    try {
+      const result = await request(requestOptions);
+      if (result.nodes) {
+         const activeNodes = [];
+         for (let i = 0; i < result.nodes.length; i++) {
+            if (result.nodes[i].online === true) {
+               activeNodes.push(
+                  {
+                     value: result.nodes[i].url + ':' + result.nodes[i].port.toString(),
+                     label: result.nodes[i].url + ':' + result.nodes[i].port.toString()
+                  });
+            }
+         }
+         log.debug(
+            `Get Total Online nodes: ${result.nodes.length}`
+         );
+         this.daemons = activeNodes;
+      }
+    } catch (err) {
+      log.debug(`Failed to get node list from API: : \n`, err);
+    }
+  };
+
+  getTimeZone = async (selectedTimeZone: string) => {
+    return selectedTimeZone;
   };
 
   setSyncStatus(syncStatus: number[]) {
